Loanmanager/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.3
django-storages[azure]